Split the different Tokenizer for the TurtleParser into different files.
* This cleans up the software architecture.
* Also equivalently refactor some of the TokenizerTests to use less RAM during compilation
  (The CTRE module is very RAM intensive)
