+++ src/parser/TurtleParser.cpp
@@ -13,7 +13,7 @@
 template <class T>
 bool TurtleParser<T>::statement() {
   _tok.skipWhitespaceAndComments();
-  return directive() || (triples() && skip<TokId::Dot>());
+  return directive() || (triples() && skip<TurtleTokenId::Dot>());
 }
 
 // ______________________________________________________________
@@ -25,8 +25,9 @@ bool TurtleParser<T>::directive() {
 // ________________________________________________________________
 template <class T>
 bool TurtleParser<T>::prefixID() {
-  if (skip<TokId::TurtlePrefix>()) {
-    if (check(pnameNS()) && check(iriref()) && check(skip<TokId::Dot>())) {
+  if (skip<TurtleTokenId::TurtlePrefix>()) {
+    if (check(pnameNS()) && check(iriref()) &&
+        check(skip<TurtleTokenId::Dot>())) {
       // strip  the angled brackes <bla> -> bla
       _prefixMap[_activePrefix] =
           _lastParseResult.substr(1, _lastParseResult.size() - 2);
@@ -42,7 +43,7 @@ bool TurtleParser<T>::prefixID() {
 // ________________________________________________________________
 template <class T>
 bool TurtleParser<T>::base() {
-  if (skip<TokId::TurtleBase>()) {
+  if (skip<TurtleTokenId::TurtleBase>()) {
     if (iriref()) {
       _baseIRI = _lastParseResult;
       return true;
@@ -57,7 +58,7 @@ bool TurtleParser<T>::base() {
 // ________________________________________________________________
 template <class T>
 bool TurtleParser<T>::sparqlPrefix() {
-  if (skip<TokId::SparqlPrefix>()) {
+  if (skip<TurtleTokenId::SparqlPrefix>()) {
     if (pnameNS() && iriref()) {
       _prefixMap[_activePrefix] = _lastParseResult;
       return true;
@@ -72,7 +73,7 @@ bool TurtleParser<T>::sparqlPrefix() {
 // ________________________________________________________________
 template <class T>
 bool TurtleParser<T>::sparqlBase() {
-  if (skip<TokId::SparqlBase>()) {
+  if (skip<TurtleTokenId::SparqlBase>()) {
     if (iriref()) {
       _baseIRI = _lastParseResult;
       return true;
@@ -108,7 +109,7 @@ bool TurtleParser<T>::triples() {
 template <class T>
 bool TurtleParser<T>::predicateObjectList() {
   if (verb() && check(objectList())) {
-    while (skip<TokId::Semicolon>()) {
+    while (skip<TurtleTokenId::Semicolon>()) {
       if (verb() && check(objectList())) {
         continue;
       }
@@ -123,7 +124,7 @@ bool TurtleParser<T>::predicateObjectList() {
 template <class T>
 bool TurtleParser<T>::objectList() {
   if (object()) {
-    while (skip<TokId::Comma>() && check(object())) {
+    while (skip<TurtleTokenId::Comma>() && check(object())) {
       continue;
     }
     return true;
@@ -142,7 +143,8 @@ bool TurtleParser<T>::verb() {
 template <class T>
 bool TurtleParser<T>::predicateSpecialA() {
   _tok.skipWhitespaceAndComments();
-  if (auto [success, word] = _tok.template getNextToken<TokId::A>(); success) {
+  if (auto [success, word] = _tok.template getNextToken<TurtleTokenId::A>();
+      success) {
     (void)word;
     _activePredicate = "<http://www.w3.org/1999/02/22-rdf-syntax-ns#type>"s;
     return true;
@@ -198,7 +200,7 @@ bool TurtleParser<T>::literal() {
 // _____________________________________________________________________
 template <class T>
 bool TurtleParser<T>::blankNodePropertyList() {
-  if (!skip<TokId::OpenSquared>()) {
+  if (!skip<TurtleTokenId::OpenSquared>()) {
     return false;
   }
   // save subject and predicate
@@ -211,7 +213,7 @@ bool TurtleParser<T>::blankNodePropertyList() {
   // the following triples have the blank node as subject
   _activeSubject = blank;
   check(predicateObjectList());
-  check(skip<TokId::CloseSquared>());
+  check(skip<TurtleTokenId::CloseSquared>());
   // restore subject and predicate
   _activeSubject = savedSubject;
   _activePredicate = savedPredicate;
@@ -221,7 +223,7 @@ bool TurtleParser<T>::blankNodePropertyList() {
 // _____________________________________________________________________
 template <class T>
 bool TurtleParser<T>::collection() {
-  if (!skip<TokId::OpenRound>()) {
+  if (!skip<TurtleTokenId::OpenRound>()) {
     return false;
   }
   raise("We do not know how to handle collections in QLever yet\n");
@@ -247,7 +249,7 @@ bool TurtleParser<T>::rdfLiteral() {
     return true;
     // TODO<joka921> this allows spaces here since the ^^ is unique in the
     // sparql syntax. is this correct?
-  } else if (skip<TokId::DoubleCircumflex>() && check(iri())) {
+  } else if (skip<TurtleTokenId::DoubleCircumflex>() && check(iri())) {
     _lastParseResult = literalString + "^^" + _lastParseResult;
     return true;
   } else {
@@ -259,7 +261,8 @@ bool TurtleParser<T>::rdfLiteral() {
 // ______________________________________________________________________
 template <class T>
 bool TurtleParser<T>::booleanLiteral() {
-  return parseTerminal<TokId::True>() || parseTerminal<TokId::False>();
+  return parseTerminal<TurtleTokenId::True>() ||
+         parseTerminal<TurtleTokenId::False>();
 }
 
 // ______________________________________________________________________
@@ -330,7 +333,7 @@ bool TurtleParser<T>::prefixedName() {
     if (!pnameNS()) {
       return false;
     }
-    parseTerminal<TokId::PnLocal, false>();
+    parseTerminal<TurtleTokenId::PnLocal, false>();
   }
   _lastParseResult = '<' + expandPrefix(_activePrefix) +
                      RdfEscaping::unescapePrefixedIri(_lastParseResult) + '>';
@@ -345,7 +348,7 @@ bool TurtleParser<T>::blankNode() {
 
 // _______________________________________________________________________
 template <class T>
-template <TokId terminal, bool SkipWhitespaceBefore>
+template <TurtleTokenId terminal, bool SkipWhitespaceBefore>
 bool TurtleParser<T>::parseTerminal() {
   if constexpr (SkipWhitespaceBefore) {
     _tok.skipWhitespaceAndComments();
@@ -363,13 +366,13 @@ bool TurtleParser<T>::parseTerminal() {
 // ________________________________________________________________________
 template <class T>
 bool TurtleParser<T>::blankNodeLabel() {
-  return parseTerminal<TokId::BlankNodeLabel>();
+  return parseTerminal<TurtleTokenId::BlankNodeLabel>();
 }
 
 // __________________________________________________________________________
 template <class T>
 bool TurtleParser<T>::pnameNS() {
-  if (parseTerminal<TokId::PnameNS>()) {
+  if (parseTerminal<TurtleTokenId::PnameNS>()) {
     // this also includes a ":" which we do not need, hence the "-1"
     _activePrefix = _lastParseResult.substr(0, _lastParseResult.size() - 1);
     _lastParseResult = "";
@@ -430,7 +433,7 @@ bool TurtleParser<T>::iriref() {
       return false;
     }
   } else {
-    if (!parseTerminal<TokId::Iriref>()) {
+    if (!parseTerminal<TurtleTokenId::Iriref>()) {
       return false;
     }
     _lastParseResult = RdfEscaping::unescapeIriref(_lastParseResult);
